requirements.txt
streamlit>=1.28.0
requests
python-dotenv